+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1357** : Fixed dragging tabs in dashboard with hidden panes.
+
 * Issue **#1368** : Fixed FindReplaceFilter as it wasn't working when used in conjunction with Data Splitter.
 
 * Issue **#1361** : Changed the way headers are parsed for the HttpCall XSLT function.